fix: tweak the mouse position after having scolled through the TestDriver
At least on WSL, using xdotool to mousemove to the current mouse
position (with --sync and all relative to a single given window) takes
about 15 seconds. The xdotool manpage claims that if no move is needed,
there won't be a delay but it looks like we can't rely on that.

tdd: refactor